tools/witch-trials-generator/requirements.txt
requests>=2.31
jinja2>=3.1
psycopg2-binary>=2.9
python-dotenv>=1.0
